+++ scripts/run_dataset.jl
@@ -1,21 +1,5 @@
 #!/usr/bin/env julia
 
-const DEFAULT_HPRLP_ENVIRONMENT = (
-    "HPRLP_ENABLE_COMPRESSIBLE_MEMORY" => "1",
-    "HPRLP_USE_ROW_REDUCTION" => "1",
-    "HPRLP_USE_ROW_COMPRESSED_AUTOTUNE" => "1",
-    "HPRLP_USE_REDUCED_COMPRESSED_AUTOTUNE" => "1",
-    "HPRLP_DEFER_REDUCED_EMPTY_ROWS_TO_CHECK" => "1",
-    "HPRLP_USE_REDUCED_NONEMPTY_CUSPARSE" => "1",
-    "HPRLP_REDUCED_RESET_MASK_ON_RESTART" => "1",
-    "HPRLP_REDUCED_RESTART_MASK_MIN_RECOVERY" => "0.25",
-    "HPRLP_REDUCED_RESTART_MASK_MIN_SAVED_COLUMNS" => "25000",
-    "HPRLP_REDUCED_RESTART_MASK_MIN_CURRENT_COLUMNS" => "95000",
-)
-for (name, value) in DEFAULT_HPRLP_ENVIRONMENT
-    get!(ENV, name, value)
-end
-
 import Pkg
 Pkg.activate(joinpath(@__DIR__, "..", "bindings", "julia", "package"); io=devnull)